Streamline entries in other-changes
This commit streamlines the entries in other changes.

I noticed quite a few misplacements of the term "now". AFAIK it may only
be written at the start or end of a sentence. In these contributions it
was usually places after the noun.

Furthermore:
- "version 1.0" -> "v1.0"
- "See #1312161 for more details" -> "Refer to PR #1312161 …"
- "The module `services.blah`" -> "The `services.blah` module"

These are all suggestions and I am open to improvements.